requirements.txt
python-dotenv
qdrant_client
matplotlib
requests
tqdm
torch
openai
transformers
sentence_transformers
huggingface_hub
ftfy
regex
tqdm
scikit-image
qdrant-client
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33

llama-index
llama-index-llms-openai
llama-index-multi-modal-llms-openai